Bullet Damage
- Bullets now interact with health script
- Heal modifier is now a float

+++ DrBounce-UnityProject/Assets/Scripts/HealthPackObject.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public class HealthPackObject : ScriptableObject
 {
     public int Health = 50;
-    public int healModifier = 2;
+    public float healModifier = 2;
 }
